Update spec language for "external" interfaces.
Engines now MUST provide a mechanism for executing commands from outside a handler.
Without such a feature there is no way to cause change within any domain.

External event recording is remains truly optional, as this is required only when
recording "integration" events, which is not a requirement of every domain.